backend/python/requirements.txt
flask==3.0.0
flask-cors==4.0.0
tensorflow==2.15.0
pillow==10.1.0
pytesseract==0.3.10
numpy==1.24.3
requests==2.31.0

